requirements.txt
astropy==1.3
cycler==0.10.0
matplotlib==2.0.0
numpy==1.12.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
six==1.10.0
